Split SVGKit into multiple files
