+++ packages/noya-svgkit/src/JSMaskFilter.ts
@@ -1,0 +1,12 @@
+import { BlurStyle, MaskFilter } from 'canvaskit';
+import { JSEmbindObject } from './Embind';
+
+export class JSMaskFilter extends JSEmbindObject implements MaskFilter {
+  static MakeBlur(
+    style: BlurStyle,
+    sigma: number,
+    respectCTM: boolean,
+  ): MaskFilter {
+    return new JSMaskFilter();
+  }
+}
